Project setup.
