+++ Home work 11./Resources/Config/Info.plist
