+++ Home work 11./Sources/App/AppDelegate.swift
@@ -1,0 +1,18 @@
+//
+//  AppDelegate.swift
+//  Home work 11.
+//
+//  Created by Ренат Сафин on 20.12.2021.
+//
+
+import UIKit
+
+@main
+class AppDelegate: UIResponder, UIApplicationDelegate {
+
+    func application(_ application: UIApplication, didFinishLaunchingWithOptions launchOptions: [UIApplication.LaunchOptionsKey: Any]?) -> Bool {
+        // Override point for customization after application launch.
+        return true
+    }
+}
+
