+++ Home work 11./Sources/App/SceneDelegate.swift
@@ -1,0 +1,18 @@
+//
+//  SceneDelegate.swift
+//  Home work 11.
+//
+//  Created by Ренат Сафин on 20.12.2021.
+//
+
+import UIKit
+
+class SceneDelegate: UIResponder, UIWindowSceneDelegate {
+
+    var window: UIWindow?
+
+    func scene(_ scene: UIScene, willConnectTo session: UISceneSession, options connectionOptions: UIScene.ConnectionOptions) {
+        guard let _ = (scene as? UIWindowScene) else { return }
+    }
+}
+
