+++ Home work 11./Sources/Scenes/TimerView.swift
@@ -1,0 +1,40 @@
+//
+//  ViewController.swift
+//  Home work 11.
+//
+//  Created by Ренат Сафин on 20.12.2021.
+//
+
+import UIKit
+
+class TimerView: UIViewController {
+
+    //MARK: - Lifecycle
+    
+    override func viewDidLoad() {
+        super.viewDidLoad()
+        // Do any additional setup after loading the view.
+        setupHierarchy()
+        setupLayout()
+        setupView()
+    }
+
+    //MARK: - Settings
+    
+    private func setupHierarchy() {
+        
+        
+    }
+    
+    private func setupLayout() {
+        
+        
+    }
+    
+    private func setupView() {
+        
+        
+    }
+
+}
+
