--- Home work 11./ViewController.swift
@@ -1,19 +1,0 @@
-//
-//  ViewController.swift
-//  Home work 11.
-//
-//  Created by Ренат Сафин on 20.12.2021.
-//
-
-import UIKit
-
-class ViewController: UIViewController {
-
-    override func viewDidLoad() {
-        super.viewDidLoad()
-        // Do any additional setup after loading the view.
-    }
-
-
-}
-
